app/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
google-auth==2.9.0
google-auth-oauthlib==0.5.2
greenlet==1.1.2
gspread==5.4.0
httplib2==0.20.4
idna==3.3
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
urllib3==1.26.9
xmltodict==0.13.0
